Properly deal with ignored paths in #getTemplateInput
The previous "solution" didn't deal with the fact that _when_ an ignore
path is not present in the input data, `applyPatch` will actually return
`nil` and not the new state (as it would if the patch was successful).

Now we instead assign the output data to a new variable, and only if
it's actually something useful (i.e. not nil) we'll eventually return
it.